requirements.txt
dask
duckdb
distributed
redis
boto3
psycopg2-binary
apache-airflow-providers-amazon[s3fs]
s3fs
sqlalchemy-redshift
apache-airflow-providers-redis
apache-airflow-providers-common-sql
apache-airflow-providers-microsoft-mssql
asyncpg
faker
pyodbc
pandas
pendulum
sqlalchemy
pyspark==3.5.0

# Additional dependencies for Spark
py4j